+++ tests/keras/applications/applications_test.py
@@ -1,5 +1,8 @@
 import pytest
 from keras.utils.test_utils import keras_test
+from keras.utils.test_utils import layer_test
+from keras.utils.generic_utils import CustomObjectScope
+from keras.models import Sequential
 from keras import applications
 from keras import backend as K
 
@@ -188,6 +191,51 @@ def test_mobilenet_no_top():
 def test_mobilenet_pooling():
     model = applications.MobileNet(weights=None, include_top=False, pooling='avg')
     assert model.output_shape == (None, 1024)
+
+
+@pytest.mark.skipif(K.backend() != 'tensorflow', reason='Requires TF backend')
+@keras_test
+def test_depthwise_conv_2d():
+    _convolution_paddings = ['valid', 'same']
+    num_samples = 2
+    stack_size = 3
+    num_row = 7
+    num_col = 6
+
+    with CustomObjectScope({'relu6': applications.mobilenet.relu6,
+                            'DepthwiseConv2D': applications.mobilenet.DepthwiseConv2D}):
+        for padding in _convolution_paddings:
+            for strides in [(1, 1), (2, 2)]:
+                for multiplier in [1, 2]:
+                    if padding == 'same' and strides != (1, 1):
+                        continue
+
+                    layer_test(applications.mobilenet.DepthwiseConv2D,
+                               kwargs={'kernel_size': (3, 3),
+                                       'padding': padding,
+                                       'strides': strides,
+                                       'depth_multiplier': multiplier},
+                               input_shape=(num_samples, num_row, num_col, stack_size))
+
+        layer_test(applications.mobilenet.DepthwiseConv2D,
+                   kwargs={'kernel_size': 3,
+                           'padding': padding,
+                           'data_format': 'channels_first',
+                           'activation': None,
+                           'depthwise_regularizer': 'l2',
+                           'bias_regularizer': 'l2',
+                           'activity_regularizer': 'l2',
+                           'depthwise_constraint': 'unit_norm',
+                           'strides': strides,
+                           'depth_multiplier': multiplier},
+                   input_shape=(num_samples, stack_size, num_row, num_col))
+
+        # Test invalid use case
+        with pytest.raises(ValueError):
+            model = Sequential([applications.mobilenet.DepthwiseConv2D(kernel_size=3,
+                                                                       padding=padding,
+                                                                       batch_input_shape=(None, None, 5, None))])
+
 
 if __name__ == '__main__':
     pytest.main([__file__])
